rest: Isolate Code-Hex/vz code from the rest code
Code-Hex/vz requires cgo/native builds, so it's preferrable to separate
it from the client rest code.
This PR achieves this by introducing interfaces to help move the
mac-specific code to a separate package
